+++ babel.config.cjs
@@ -1,0 +1,13 @@
+module.exports = {
+  presets: [
+    "@babel/preset-env",
+    "@babel/preset-react",
+    "@babel/preset-typescript",
+  ],
+  plugins: [
+    "@babel/plugin-proposal-class-properties",
+    "@babel/plugin-syntax-jsx",
+    "css-modules-transform",
+    "inline-react-svg",
+  ],
+};